+++ extensions/applicationinsights-angularplugin-js/src/AngularPlugin.ts
@@ -4,22 +4,27 @@
  */
 
 import {
-    IConfig, IPageViewTelemetry, IMetricTelemetry, IAppInsights
+    IConfig, IPageViewTelemetry, IMetricTelemetry, IAppInsights, PropertiesPluginIdentifier, Util
 } from "@microsoft/applicationinsights-common";
 import {
     IPlugin, IConfiguration, IAppInsightsCore,
     ITelemetryPlugin, BaseTelemetryPlugin, CoreUtils, ITelemetryItem, ITelemetryPluginChain,
-    IProcessTelemetryContext, _InternalMessageId, LoggingSeverity, ICustomProperties
+    IProcessTelemetryContext, _InternalMessageId, LoggingSeverity, ICustomProperties, getLocation
 } from "@microsoft/applicationinsights-core-js";
 import { IAngularExtensionConfig } from './Interfaces/IAngularExtensionConfig';
+// For types only
+import * as properties from "@microsoft/applicationinsights-properties-js";
+
+const NAVIGATIONEND = "NavigationEnd";
 
 export default class AngularPlugin extends BaseTelemetryPlugin {
     public priority = 186;
     public identifier = 'AngularPlugin';
 
     private _analyticsPlugin: IAppInsights;
+    private _propertiesPlugin: properties.PropertiesPlugin;
 
-    initialize(config: IConfiguration & IConfig, core: IAppInsightsCore, extensions: IPlugin[], pluginChain?:ITelemetryPluginChain) {
+    initialize(config: IConfiguration & IConfig, core: IAppInsightsCore, extensions: IPlugin[], pluginChain?: ITelemetryPluginChain) {
         super.initialize(config, core, extensions, pluginChain);
         let ctx = this._getTelCtx();
         let extConfig = ctx.getExtCfg<IAngularExtensionConfig>(this.identifier, { router: null });
@@ -28,21 +33,29 @@ export default class AngularPlugin extends BaseTelemetryPlugin {
             if (identifier === 'ApplicationInsightsAnalytics') {
                 this._analyticsPlugin = (ext as any) as IAppInsights;
             }
+            if (identifier === PropertiesPluginIdentifier) {
+                this._propertiesPlugin = (ext as any) as properties.PropertiesPlugin;
+            }
         });
         if (extConfig.router) {
+            let isPageInitialLoad = true;
+            if (isPageInitialLoad) {
+                const pageViewTelemetry: IPageViewTelemetry = {
+                    uri: extConfig.router.url
+                };
+                this.trackPageView(pageViewTelemetry);
+            }
             extConfig.router.events.subscribe(event => {
-                if (event.constructor.name === "NavigationEnd") {
-                    // Timeout to ensure any changes to the DOM made by route changes get included in pageView telemetry
-                    setTimeout(() => {
-                        const pageViewTelemetry: IPageViewTelemetry = { uri: extConfig.router.url };
-                        this.trackPageView(pageViewTelemetry);
-                    }, 500);
+                if (event.constructor.name === NAVIGATIONEND) {
+                    // for page initial load, do not call trackPageView twice
+                    if (isPageInitialLoad) {
+                        isPageInitialLoad = false;
+                        return;
+                    }
+                    const pageViewTelemetry: IPageViewTelemetry = { uri: extConfig.router.url };
+                    this.trackPageView(pageViewTelemetry);
                 }
             });
-            const pageViewTelemetry: IPageViewTelemetry = {
-                uri: extConfig.router.url
-            };
-            this.trackPageView(pageViewTelemetry);
         }
     }
 
@@ -65,6 +78,11 @@ export default class AngularPlugin extends BaseTelemetryPlugin {
 
     trackPageView(pageView: IPageViewTelemetry) {
         if (this._analyticsPlugin) {
+            const location = getLocation();
+            if (this._propertiesPlugin && this._propertiesPlugin.context && this._propertiesPlugin.context.telemetryTrace) {
+                this._propertiesPlugin.context.telemetryTrace.traceID = Util.generateW3CId();
+                this._propertiesPlugin.context.telemetryTrace.name = location && location.pathname || "_unknown_";
+            }
             this._analyticsPlugin.trackPageView(pageView);
         } else {
             this.diagLog().throwInternal(